Client Folder Structure
