--- client/src/ClientProfilePage.css
@@ -1,9 +1,0 @@
-.select_box{
-    margin-bottom: 20px;
-}
-
-.profileh1{
-    
-    text-align: center;
-    margin-bottom: 15px;
-}
