--- client/src/SignUpPage.css
@@ -1,44 +1,0 @@
-.signup_form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    background-color: #f2f2f2;
-    padding: 30px;
-    border-radius: 10px;
-    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
-    width: 50%;
-    margin: auto;
-  }
-  
-  .signup_form h2 {
-    margin-bottom: 30px;
-  }
-  
-  .signup_label_username,.signup_label_pass,.signup_label_conpass {
-    display: flex;
-    margin-top: 10px;
-    margin-bottom: 5px;
-    align-items: center;
-  }
-  
-  .signup_input {
-    width: 100%;
-    padding: 10px;
-    border-radius: 5px;
-    border: none;
-    margin-bottom: 20px;
-  }
-  
-  .signup_button {
-    background-color: #4CAF50;
-    color: white;
-    padding: 10px 20px;
-    border-radius: 5px;
-    border: none;
-    cursor: pointer;
-  }
-  
-  .signup_button:hover {
-    background-color: #3e8e41;
-  }
-
