--- client/src/clients/pages/welcome.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-welcome.propTypes = {
-    
-};
-
-function welcome(props) {
-    return (
-        <div>
-            welcome
-        </div>
-    );
-}
-
-export default welcome;
