--- client/src/guests/pages/WelcomeGuest.css
@@ -1,12 +1,0 @@
-.container {
-    display: flex;
-    width: 100%;
-    height: 100%;
-    align-items: center;
-    justify-content: center;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: red;
-}
