--- client/src/guests/pages/WelcomeGuest.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import './WelcomeGuest.css';
-
-WelcomeGuest.propTypes = {
-    
-};
-
-function WelcomeGuest(props) {
-    return (
-        <div className='container'>
-            <h1>Group 9 Project</h1>
-        </div>
-    );
-}
-
-export default WelcomeGuest;
