+++ client/src/index.js
@@ -1,21 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import WelcomeGuest from './guests/pages/WelcomeGuest';
 import reportWebVitals from './reportWebVitals';
 import App from './App';
 import './index.css';
-// import App from './App';
-// import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-  <WelcomeGuest/>
-   <App/>    
+    <App/>    
   </React.StrictMode>
 );
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
