+++ client/src/pages/ClientPage.js
@@ -19,4 +19,8 @@ function ClientPage() {
   );
 }
 
+<style>
+
+</style>
+
 export default ClientPage;
