+++ client/src/pages/index.js
@@ -1,0 +1,13 @@
+import HomePage from './HomePage';
+import ClientPage from './ClientPage';
+import ClientProfilePage from './ClientProfilePage';
+import LoginPage from './LoginPage';
+import SignUpPage from './SignUpPage';
+
+export { 
+  HomePage, 
+  ClientPage,
+  ClientProfilePage,
+  LoginPage,
+  SignUpPage,
+};
